examples: move hello message structure into common/hello.h
