+++ examples/03-read-to-persistent/client.c
@@ -19,20 +19,7 @@
 #endif /* USE_LIBPMEM */
 
 #include "common-conn.h"
-
-enum lang_t {en, es};
-
-static const char *hello_str[] = {
-	[en] = "Hello world!",
-	[es] = "¡Hola Mundo!"
-};
-
-#define LANG_NUM	(sizeof(hello_str) / sizeof(hello_str[0]))
-
-struct hello_t {
-	enum lang_t lang;
-	char str[KILOBYTE];
-};
+#include "hello.h"
 
 static inline void
 write_hello_str(struct hello_t *hello, enum lang_t lang)
