+++ examples/09-flush-to-persistent-GPSPM/client.c
@@ -24,20 +24,7 @@
 
 /* Generated by the protocol buffer compiler from: GPSPM_flush.proto */
 #include "GPSPM_flush.pb-c.h"
-
-enum lang_t {en, es};
-
-static const char *hello_str[] = {
-	[en] = "Hello world!",
-	[es] = "¡Hola Mundo!"
-};
-
-#define LANG_NUM	(sizeof(hello_str) / sizeof(hello_str[0]))
-
-struct hello_t {
-	enum lang_t lang;
-	char str[KILOBYTE];
-};
+#include "hello.h"
 
 #define FLUSH_ID	(void *)0xF01D /* a random identifier */
